+++ Controller/SmtpsController.php
@@ -160,12 +160,21 @@ class SmtpsController extends AppController {
 	
 		$result = false;
 		$message = __('Errore durante l\'operazione.');
-		
-		switch($this->request->data['Smtp']['action']) {
-			case 'bulkDelete':
-				list($result, $message) = $this->Smtp->bulkDelete($this->request->data['Smtp']['selected']);
-			break;
+		if(!empty($this->request->data['Smtp']['selected'])) {
+			switch($this->request->data['Smtp']['action']) {
+				case 'bulkDelete':
+					list($result, $message) = $this->Smtp->bulkDelete($this->request->data['Smtp']['selected']);
+				break;
+				default:
+					$this->Session->setFlash(__("Seleziona una operazione valida"), 'default', array(), 'error');
+					$this->redirect($this->referer(true, '/'));
+			}
 		}
+		else {
+			$this->Session->setFlash(__("Seleziona almeno un indirizzo"), 'default', array(), 'error');
+			$this->redirect($this->referer(true, '/'));
+		}
+		
 		if($result) {
 			if($message === true) {
 				$message = __('Operazione eseguita con successo.');
@@ -215,22 +224,20 @@ class SmtpsController extends AppController {
 	
 	protected function __securitySettings_bulk() {
 		$this->request->onlyAllow('post');
-		if(isset($this->request->data['Smtp']['action']) && isset($this->request->data['Smtp']['selected'])) {
-			switch($this->request->data['Smtp']['action']) {
-				case 'bulkDelete':
-					$this->Security->allowedControllers = array('smtps');
-					$this->Security->allowedActions = array('index');
-					$this->request->data['Smtp']['selected'] = explode(',', $this->request->data['Smtp']['selected']);
-					foreach($this->request->data['Smtp']['selected'] as $selected) {
-						$this->Xuser->checkPerm($this->Smtp, $selected);
-					}
-				break;
-				default:
-					$this->Security->blackHole($this, 'auth');
-			}
-		}
-		else {
-			$this->Security->blackHole($this, 'auth');
+		if(!isset($this->request->data['Smtp']['action']))
+			$this->request->data['Smtp']['action'] = null;
+		if(!isset($this->request->data['Smtp']['selected']))
+			$this->request->data['Smtp']['selected'] = array();
+			
+		switch($this->request->data['Smtp']['action']) {
+			case 'bulkDelete':
+				$this->Security->allowedControllers = array('smtps');
+				$this->Security->allowedActions = array('index');
+				$this->request->data['Smtp']['selected'] = explode(',', $this->request->data['Smtp']['selected']);
+				foreach($this->request->data['Smtp']['selected'] as $selected) {
+					$this->Xuser->checkPerm($this->Smtp, $selected);
+				}
+			break;
 		}
 	}
